message type always equals to "EVENT" issue fixed
